pyproject.toml
[tool.poetry]
name = "qanoniah-rag"
version = "0.1.0"
description = ""
authors = ["Feras <firas.almanna3@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.13"

langsmith = "^0.1.93"
langgraph = "^0.1.14"
langchain = "^0.2.11"
langchain-community = "0.2.6"
langchain-ollama = "^0.1.0"
langchain-huggingface = "^0.0.3"
langchain-qdrant = "^0.1.3"
langchain-milvus = "^0.1.3"
langchain-chroma = "^0.1.2"
langchain-postgres = "^0.0.9"

python-dotenv = "^1.0.1"
semantic-text-splitter = "^0.14.1"
tqdm = "^4.66.4"
pgvecto-rs = { extras = ["sdk"], version = "^0.1.5" }

protobuf = "4.25.4"
infinity-emb = {extras = ["optimum", "torch"], version = "^0.0.53"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
